azure_hijack/requirements.txt
azure-cli-core>=2.0.55
azure-graphrbac>=0.61.1
azure-mgmt-authorization>=0.60.0